refactor: 공통 설정 모듈 이동
