+++ maeil-support/src/main/java/maeilsupport/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package maeilmail.support;
+package maeilsupport.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
