+++ maeil-wiki/src/main/java/maeilwiki/WikiConfiguration.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan
+@ComponentScan(basePackages = {"maeilwiki", "maeilsupport"})
 @EnableAutoConfiguration
 public class WikiConfiguration {
 }
